backend/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
typing-extensions>=4.12.0
pytz>=2023.3
pytest>=7.0.0
